+++ app/components/MyButton.js
@@ -22,8 +22,8 @@ import {
 } from 'react-native-fbsdk';
 
 import MainView from '../views/Home.js';
-import ImadaEventView from '../ImadaEventView.js';
-import {mainTextColor, darkBlue} from '../constants.js';
+import ImadaEventView from '../../ImadaEventView.js';
+import {mainTextColor, darkBlue} from '../config/constants.js';
 
 export default class MyButton extends Component {
     static propTypes = {
