+++ app/views/Settings.js
@@ -20,8 +20,8 @@ import {
 
 import MyButton from '../components/MyButton.js';
 import MainView from './Home.js';
-import ImadaEventView from '../ImadaEventView.js';
-import {mainTextColor, darkBlue} from '../constants.js';
+import ImadaEventView from '../../ImadaEventView.js';
+import {mainTextColor, darkBlue} from '../config/constants.js';
 
 
 export default class Settings extends Component {
